Refactor dogo structure to include agents package
agents package is used to provide a configuration for multiple bot agents (such as Docker, Lambda,
Kubernetes Jobs).
